text-recognition/requirements.txt
opencv-python==4.6.0.66
numpy<2.0
vietocr==0.3.13
python-dotenv==1.1.0
